fix: Don't delete merged chapters if local chapters are enabled

+++ app/src/main/java/eu/kanade/tachiyomi/ui/manga/MangaDetailPresenter.kt
@@ -648,14 +648,15 @@ class MangaDetailPresenter(
         presenterScope.launchIO {
             db.deleteMergeMangaForType(mangaId, mergeType).executeAsBlocking()
             updateMangaFlow()
-
-            val mergedChapters =
-                db.getChapters(currentManga()).executeOnIO().filter {
+            val (mergedChapters, nonMergedChapters) =
+                db.getChapters(currentManga()).executeOnIO().partition {
                     it.isMergedChapterOfType(mergeType)
                 }
-
-            downloadManager.deleteChapters(mergedChapters, currentManga())
+            if (!libraryPreferences.enableLocalChapters().get()) {
+                downloadManager.deleteChapters(mergedChapters, currentManga())
+            }
             db.deleteChapters(mergedChapters).executeOnIO()
+
             updateAllFlows()
         }
     }
